Small adjustments. Enable data cache for DSP mem/AXIWRAM on ARM9.

+++ source/arm9/ipc_handler.c
@@ -85,7 +85,7 @@ u32 IPC_handleCmd(u8 cmdId, u8 inBufs, u8 outBufs, const u32 *const buf)
 			panic();
 	}
 
-	for(u32 i = inBufs; i < inBufs + outBufs; i++)
+	for(u32 i = inBufs; i < (u32)inBufs + outBufs; i++)
 	{
 		const IpcBuffer *const outBuf = (IpcBuffer*)&buf[i * sizeof(IpcBuffer) / 4];
 		flushDCacheRange(outBuf->ptr, outBuf->size);
